Add Automatic Commits With Github Actions
Created a GitHub Action, that is triggered manually,
that commits generated commits to a destined branch.

**Usage**

1. Fork the repository (this is a must).
1. Navigate to the "Actions" tab in GitHub,
   and select the action named: "Git Commits As A Service",
   and select the intervals (in minutes) you wish create a commit.
   > The default is `60` minutes.
1. Then run the workflow.

**Flow Explanation**

The workflow fetches the repository to its working directory.
It views who triggered the workflow (that is why you should own the fork),
and sees your "git user.name" and searches for your "git user.email".
Then, it creates a new branch called "service", switches to it,
and creates a file named "commits.txt" on the root directory.
Now, it creates a commit for every interval (in minutes, stated above),
that its content is the "current date" (this way every commit is unique,
and won't clash with the previous commits), and types it to the
"commits.txt" file, and pushes it to the "origin/service" branch.

In case you wish to delete all of the commits at once,
you can delete the "origin/service" branch:

```
git push origin --delete service
```

Please take note, that if you wish to cancel the workflow run,
it may take about 2 minutes for it to cancel, probably because
it is very busy, but don't panic if it does not cancel immediately.

Signed-off-by: Tal Jacob <taljacob2@gmail.com>